reworked the hero section of the website. added hero bg image
